test classes are merged into one class
